# solution/MainServer/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IumTweb_project",
    "description": "This API is part of the IUM-TWEB project for the 2023-2024 academic year at the Dipartimento di Informatica, Università di Torino. Designed to enable both fans and pundits to access and analyze a wide range of football data, this API supports functionalities across multiple servers (two Express servers and one Java Spring server). It provides interfaces for querying football competitions, clubs, and player statistics, and integrates with MongoDB and Postgres databases to manage dynamic and static datasets respectively. [ IUM-TWEB Group Accornero, Aiello, Bonincontro ]",
    "version": "1.0.0",
    "contact": {
      "name": "IUM-TWEB Group Accornero, Aiello, Bonincontro"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "MainServer - Can handle all API requests."
    },
    {
      "url": "http://localhost:3001",
      "description": "ExpressServer"
    },
    {
      "url": "http://localhost:8081",
      "description": "SpringServer"
    }
  ],
  "paths": {
...
    "/gamelineups/getPlayerNumberByIdPlayer/{idPlayer}": {
      "get": {
        "tags": [
          "Game Lineups"
        ],
        "summary": "Get player number by player ID",
        "description": "Retrieves the player number assigned to a player for a specific game based on the player's ID.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:3001",
            "description": "ExpressServer"
          }
        ],
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the player.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - Returns the player number.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playerNumber": {
                      "type": "integer",
                      "description": "Number assigned to the player for the game."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/gamelineups/getMatchPlayers": {
      "get": {
        "tags": [
          "Game Lineups"
        ],
        "summary": "Retrieve players participating in a match",
        "description": "Fetches a list of all players participating in a specific match, categorized by home and away teams.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:3001",
            "description": "ExpressServer"
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "query",
            "required": true,
            "description": "The game ID for which player information is needed.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "home_club_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the home club.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "away_club_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the away club.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - Returns a list of players for both teams, divided in starting_lineup and substitutes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "home_lineup": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "game_id": {
                              "type": "integer"
                            },
                            "club_id": {
                              "type": "integer"
                            },
                            "lineup":{
                              "type": "object",
                              "properties": {
                                "starting_lineup": {
                                  "type": "object",
                                  "properties": {
                                    "player_id": {
                                      "type": "integer",
                                      "example": 8192
                                    },
                                    "player_name": {
                                      "type": "string",
                                      "example": "Cristiano Ronaldo"
                                    },
                                    "position": {
                                      "type": "string",
                                      "example": "Left-Forward"
                                    },
                                    "team_captain": {
                                      "type": "integer",
                                      "example": 1
                                    },
                                    "number": {
                                      "type": "integer",
                                      "example": 7
                                    }
                                  }
                                },
                                "substitutes": {
                                  "type": "object",
                                  "properties": {
                                    "player_id": {
                                      "type": "integer",
                                      "example": 8192
                                    },
                                    "player_name": {
                                      "type": "string",
                                      "example": "Cristiano Ronaldo"
                                    },
                                    "position": {
                                      "type": "string",
                                      "example": "Left-Forward"
                                    },
                                    "team_captain": {
                                      "type": "integer",
                                      "example": 1
                                    },
                                    "number": {
                                      "type": "integer",
                                      "example": 7
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": " :( - Error Message"
          },
          "internal_code": {
            "type": "string",
            "description": " :( - Code Error"
          }
        }
      }
    }
  }
}